refactor : 내 정보 변경 부분 controller 병합 (#196)
